int phone number check to double

+++ MedixCollege/Controllers/ExternalController.cs
@@ -464,9 +464,9 @@ namespace MedixCollege.Controllers
 
             if (fc["Telephone"] != null)
             {
-                int phoneNumber = 0;
+                double phoneNumber = 0;
 
-                int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+                double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
                 if (phoneNumber == 0)
                 {
@@ -616,9 +616,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
